♻️ Extract Default function
