+++ default.go
@@ -1,0 +1,9 @@
+package requestid
+
+func Default() *Options {
+	return &Options{
+		generator:      defaultGenerator,
+		requestHeader:  "X-Request-ID",
+		responseHeader: "X-Request-ID",
+	}
+}
